Transform `chain.ts` file to `lib/contracts` shared lib
We just moved `bitcoin.ts` to `lib/bitcoin`. Here we do the same
transition for `chain.ts`.
