+++ typescript/src/lib/contracts/chain-identifier.ts
@@ -1,0 +1,15 @@
+/**
+ * Represents a generic chain identifier.
+ */
+export interface Identifier {
+  /**
+   * Identifier as an un-prefixed hex string.
+   */
+  identifierHex: string
+  /**
+   * Checks if two identifiers are equal.
+   *
+   * @param identifier Another identifier
+   */
+  equals(identifier: Identifier): boolean
+}
