+++ typescript/src/lib/contracts/index.ts
@@ -1,0 +1,6 @@
+export * from "./bridge"
+export * from "./chain-event"
+export * from "./chain-identifier"
+export * from "./tbtc-token"
+export * from "./tbtc-vault"
+export * from "./wallet-registry"
