+++ typescript/src/redemption.ts
@@ -1,21 +1,17 @@
+import bcoin from "bcoin"
 import { BigNumber } from "ethers"
 import {
-  createAddressFromPublicKey,
+  BitcoinNetwork,
+  createKeyRing,
   decomposeRawTransaction,
   RawTransaction,
   UnspentTransactionOutput,
   Client as BitcoinClient,
   TransactionHash,
-  isP2PKHScript,
-  isP2WPKHScript,
-} from "./bitcoin"
-import { Bridge, Event, Identifier, TBTCToken } from "./chain"
+} from "./lib/bitcoin"
+import { Bridge, Event, Identifier, TBTCToken } from "./lib/contracts"
 import { assembleTransactionProof } from "./proof"
 import { determineWalletMainUtxo, WalletState } from "./wallet"
-import { BitcoinNetwork, toBitcoinJsLibNetwork } from "./bitcoin-network"
-import { Psbt, Transaction } from "bitcoinjs-lib"
-import { ECPairFactory } from "ecpair"
-import * as tinysecp from "tiny-secp256k1"
 import { Hex } from "./hex"
 
 /**
@@ -144,25 +140,15 @@ export async function submitRedemptionTransaction(
     transactionHex: mainUtxoRawTransaction.transactionHex,
   }
 
-  const bitcoinNetwork = await bitcoinClient.getNetwork()
-
-  // eslint-disable-next-line new-cap
-  const walletKeyPair = ECPairFactory(tinysecp).fromWIF(
-    walletPrivateKey,
-    toBitcoinJsLibNetwork(bitcoinNetwork)
-  )
-  const walletPublicKey = walletKeyPair.publicKey.toString("hex")
-
   const redemptionRequests = await getWalletRedemptionRequests(
     bridge,
-    walletPublicKey,
+    createKeyRing(walletPrivateKey).getPublicKey().toString("hex"),
     redeemerOutputScripts,
     "pending"
   )
 
   const { transactionHash, newMainUtxo, rawTransaction } =
     await assembleRedemptionTransaction(
-      bitcoinNetwork,
       walletPrivateKey,
       mainUtxoWithRaw,
       redemptionRequests,
@@ -256,7 +242,6 @@ async function getWalletRedemptionRequests(
  *        - there is at least one redemption
  *        - the `requestedAmount` in each redemption request is greater than
  *          the sum of its `txFee` and `treasuryFee`
- * @param bitcoinNetwork - The target Bitcoin network.
  * @param walletPrivateKey - The private key of the wallet in the WIF format
  * @param mainUtxo - The main UTXO of the wallet. Must match the main UTXO held
  *        by the on-chain Bridge contract
@@ -269,7 +254,6 @@ async function getWalletRedemptionRequests(
  *          - the redemption transaction in the raw format
  */
 export async function assembleRedemptionTransaction(
-  bitcoinNetwork: BitcoinNetwork,
   walletPrivateKey: string,
   mainUtxo: UnspentTransactionOutput & RawTransaction,
   redemptionRequests: RedemptionRequest[],
@@ -283,46 +267,19 @@ export async function assembleRedemptionTransaction(
     throw new Error("There must be at least one request to redeem")
   }
 
-  const network = toBitcoinJsLibNetwork(bitcoinNetwork)
-  // eslint-disable-next-line new-cap
-  const walletKeyPair = ECPairFactory(tinysecp).fromWIF(
-    walletPrivateKey,
-    network
-  )
-  const walletAddress = createAddressFromPublicKey(
-    Hex.from(walletKeyPair.publicKey),
-    bitcoinNetwork,
-    witness
-  )
+  const walletKeyRing = createKeyRing(walletPrivateKey, witness)
+  const walletAddress = walletKeyRing.getAddress("string")
 
-  const psbt = new Psbt({ network })
-  psbt.setVersion(1)
-
-  // Add input (current main UTXO).
-  const previousOutput = Transaction.fromHex(mainUtxo.transactionHex).outs[
-    mainUtxo.outputIndex
+  // Use the main UTXO as the single transaction input
+  const inputCoins = [
+    bcoin.Coin.fromTX(
+      bcoin.MTX.fromRaw(mainUtxo.transactionHex, "hex"),
+      mainUtxo.outputIndex,
+      -1
+    ),
   ]
-  const previousOutputScript = previousOutput.script
-  const previousOutputValue = previousOutput.value
 
-  if (isP2PKHScript(previousOutputScript)) {
-    psbt.addInput({
-      hash: mainUtxo.transactionHash.reverse().toBuffer(),
-      index: mainUtxo.outputIndex,
-      nonWitnessUtxo: Buffer.from(mainUtxo.transactionHex, "hex"),
-    })
-  } else if (isP2WPKHScript(previousOutputScript)) {
-    psbt.addInput({
-      hash: mainUtxo.transactionHash.reverse().toBuffer(),
-      index: mainUtxo.outputIndex,
-      witnessUtxo: {
-        script: previousOutputScript,
-        value: previousOutputValue,
-      },
-    })
-  } else {
-    throw new Error("Unexpected main UTXO type")
-  }
+  const transaction = new bcoin.MTX()
 
   let txTotalFee = BigNumber.from(0)
   let totalOutputsValue = BigNumber.from(0)
@@ -346,34 +303,44 @@ export async function assembleRedemptionTransaction(
     //       use the proposed fee and add the difference to outputs proportionally.
     txTotalFee = txTotalFee.add(request.txMaxFee)
 
-    psbt.addOutput({
-      script: Buffer.from(request.redeemerOutputScript, "hex"),
+    transaction.addOutput({
+      script: bcoin.Script.fromRaw(
+        Buffer.from(request.redeemerOutputScript, "hex")
+      ),
       value: outputValue.toNumber(),
     })
   }
 
-  // If there is a change output, add it to the transaction.
+  // If there is a change output, add it explicitly to the transaction.
+  // If we did not add this output explicitly, the bcoin library would add it
+  // anyway during funding, but if the value of the change output was very low,
+  // the library would consider it "dust" and add it to the fee rather than
+  // create a new output.
   const changeOutputValue = mainUtxo.value
     .sub(totalOutputsValue)
     .sub(txTotalFee)
   if (changeOutputValue.gt(0)) {
-    psbt.addOutput({
-      address: walletAddress,
+    transaction.addOutput({
+      script: bcoin.Script.fromAddress(walletAddress),
       value: changeOutputValue.toNumber(),
     })
   }
 
-  psbt.signAllInputs(walletKeyPair)
-  psbt.finalizeAllInputs()
+  await transaction.fund(inputCoins, {
+    changeAddress: walletAddress,
+    hardFee: txTotalFee.toNumber(),
+    subtractFee: false,
+  })
 
-  const transaction = psbt.extractTransaction()
-  const transactionHash = TransactionHash.from(transaction.getId())
+  transaction.sign(walletKeyRing)
+
+  const transactionHash = TransactionHash.from(transaction.txid())
   // If there is a change output, it will be the new wallet's main UTXO.
   const newMainUtxo = changeOutputValue.gt(0)
     ? {
         transactionHash,
         // It was the last output added to the transaction.
-        outputIndex: transaction.outs.length - 1,
+        outputIndex: transaction.outputs.length - 1,
         value: changeOutputValue,
       }
     : undefined
@@ -382,7 +349,7 @@ export async function assembleRedemptionTransaction(
     transactionHash,
     newMainUtxo,
     rawTransaction: {
-      transactionHex: transaction.toHex(),
+      transactionHex: transaction.toRaw().toString("hex"),
     },
   }
 }
