+++ typescript/test/ethereum.test.ts
@@ -10,7 +10,7 @@ import { abi as TBTCTokenABI } from "@keep-network/tbtc-v2/artifacts/TBTC.json"
 import { abi as WalletRegistryABI } from "@keep-network/ecdsa/artifacts/WalletRegistry.json"
 import { MockProvider } from "@ethereum-waffle/provider"
 import { waffleChai } from "@ethereum-waffle/chai"
-import { TransactionHash, computeHash160 } from "../src/bitcoin"
+import { TransactionHash, computeHash160 } from "../src/lib/bitcoin"
 import { Hex } from "../src/hex"
 
 chai.use(waffleChai)
